+++ lib/compat/wordpress-7.0/class-gutenberg-rest-old-templates-controller.php
@@ -1,6 +1,8 @@
 <?php
-
-class Gutenberg_REST_Old_Templates_Controller extends WP_REST_Templates_Controller {
+/**
+ * @see Gutenberg_REST_Templates_Controller_7_0
+ */
+class Gutenberg_REST_Old_Templates_Controller extends Gutenberg_REST_Templates_Controller_7_0 {
 	public function get_template_fallback( $request ) {
 		// Check active_templates experiment status.
 		if ( ! gutenberg_is_experiment_enabled( 'active_templates' ) ) {