+++ packages/block-library/src/breadcrumbs/edit.js
@@ -30,8 +30,8 @@ export default function BreadcrumbEdit( {
 } ) {
 	const {
 		separator,
-		showHomeLink,
-		showLastItem,
+		showHomeItem,
+		showCurrentItem,
 		prefersTaxonomy,
 		showOnHomePage,
 	} = attributes;
@@ -143,7 +143,7 @@ export default function BreadcrumbEdit( {
 		( _showTerms && ! hasTermsAssigned );
 	if ( showPlaceholder ) {
 		const placeholderItems = [];
-		if ( showHomeLink ) {
+		if ( showHomeItem ) {
 			placeholderItems.push( __( 'Home' ) );
 		}
 		if ( templateSlug && ! postId ) {
@@ -168,7 +168,7 @@ export default function BreadcrumbEdit( {
 							</a>
 						</li>
 					) ) }
-					{ showLastItem && (
+					{ showCurrentItem && (
 						<li>
 							<span aria-current="page">{ __( 'Current' ) }</span>
 						</li>
@@ -185,48 +185,48 @@ export default function BreadcrumbEdit( {
 					resetAll={ () => {
 						setAttributes( {
 							separator: separatorDefaultValue,
-							showHomeLink: true,
-							showLastItem: true,
+							showHomeItem: true,
+							showCurrentItem: true,
 						} );
 					} }
 					dropdownMenuProps={ dropdownMenuProps }
 				>
 					<ToolsPanelItem
-						label={ __( 'Show home link' ) }
+						label={ __( 'Show home breadcrumb' ) }
 						isShownByDefault
-						hasValue={ () => ! showHomeLink }
+						hasValue={ () => ! showHomeItem }
 						onDeselect={ () =>
 							setAttributes( {
-								showHomeLink: true,
+								showHomeItem: true,
 							} )
 						}
 					>
 						<ToggleControl
 							__nextHasNoMarginBottom
-							label={ __( 'Show home link' ) }
+							label={ __( 'Show home breadcrumb' ) }
 							onChange={ ( value ) =>
-								setAttributes( { showHomeLink: value } )
+								setAttributes( { showHomeItem: value } )
 							}
-							checked={ showHomeLink }
+							checked={ showHomeItem }
 						/>
 					</ToolsPanelItem>
 					<ToolsPanelItem
-						label={ __( 'Show last item' ) }
+						label={ __( 'Show current breadcrumb' ) }
 						isShownByDefault
-						hasValue={ () => ! showLastItem }
+						hasValue={ () => ! showCurrentItem }
 						onDeselect={ () =>
 							setAttributes( {
-								showLastItem: true,
+								showCurrentItem: true,
 							} )
 						}
 					>
 						<ToggleControl
 							__nextHasNoMarginBottom
-							label={ __( 'Show last item' ) }
+							label={ __( 'Show current breadcrumb' ) }
 							onChange={ ( value ) =>
-								setAttributes( { showLastItem: value } )
+								setAttributes( { showCurrentItem: value } )
 							}
-							checked={ showLastItem }
+							checked={ showCurrentItem }
 						/>
 					</ToolsPanelItem>
 					<ToolsPanelItem