+++ packages/global-styles-ui/src/font-library-modal/font-card.tsx
@@ -10,12 +10,12 @@ import {
 	FlexItem,
 } from '@wordpress/components';
 import { Icon, chevronLeft, chevronRight } from '@wordpress/icons';
+import type { FontFamily } from '@wordpress/core-data';
 
 /**
  * Internal dependencies
  */
 import FontDemo from './font-demo';
-import type { FontFamily } from './types';
 
 function FontCard( {
 	font,