+++ packages/global-styles-ui/src/font-library-modal/library-font-variant.tsx
@@ -3,6 +3,7 @@
  */
 import { useContext, useId } from '@wordpress/element';
 import { CheckboxControl, Flex } from '@wordpress/components';
+import type { FontFace, FontFamily } from '@wordpress/core-data';
 
 /**
  * Internal dependencies
@@ -10,7 +11,6 @@ import { CheckboxControl, Flex } from '@wordpress/components';
 import { getFontFaceVariantName } from './utils';
 import { FontLibraryContext } from './context';
 import FontDemo from './font-demo';
-import type { FontFace, FontFamily } from './types';
 
 function LibraryFontVariant( {
 	face,